Apaguei o model2
Faltava algumas alterações no arquivo "model_cadastro" então eu tinha criado a "model_cadastro2". Apaguei o arquivo que eu tinha criado e joguei dentro do "model_cadastro".
