--- portalsalaberga/controller_cadastro.php
@@ -1,2 +1,0 @@
-<?php 
-?>
